requirements.txt
openai>=1.40.0
anthropic>=0.34.0
azure-ai-documentintelligence>=1.0.0
azure-storage-blob>=12.19.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.2
faiss-cpu>=1.7.4
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.15
langchain-community>=0.2.0
tiktoken>=0.7.0
requests>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6

# Local PDF extraction (replaces Azure DI in UI flow)
pymupdf4llm>=0.0.17
pdfplumber>=0.11.0

# UI backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
slowapi>=0.1.9
